finish

+++ fileSystem/main.c
@@ -11,13 +11,36 @@
 typedef char* String;
 
 int main(){
-    fileAdd();  
-    removeFile();
-    fileRead();
-    modify();
-    fileRead();
-    ls();
-    makeDir();
-    rmDir();
+    char str[5] = "";
+
+    system("clear");
+    
+    while(1){
+        *str = "";
+        printf("fileSystem ~ %% ");
+        scanf("%s", str);
+
+
+        if(strstr(str, "makefile") != NULL){
+            fileAdd();
+        }else if(strstr(str, "removefile") != NULL){
+            removeFile();
+        }else if(strstr(str, "readfile") != NULL){
+            fileRead();
+        }else if(strstr(str, "modify") != NULL){
+            modify();
+        }else if(strstr(str, "readDir") != NULL){
+            ls();
+        }else if(strstr(str, "makeDir") != NULL){
+            makeDir();
+        }else if(strstr(str, "removeDir") != NULL){
+            rmDir();
+        }else if(strstr(str, "end") != NULL){
+            printf("종료");
+            return 0;
+        }else{
+            printf("not an identifier: %s\n", str); 
+        }
+    }
     return 0;
 }
